feat(debuglink): introduce debuglink_context() context manager
[no changelog]

Closes #6366.

- Replace manual DebugLink setup across the debug CLI with a single
  @contextmanager that opens the debug transport and yields a DebugLink,
  ensuring proper cleanup even if an exception occurs
- Rename `filter` -> `log_filter` in set_log_filter() (avoids shadowing
  Python builtin, ruff A002); type as `str | None` since required=False
- Docstring aligned: remove 'ready' claim (readiness is caller-managed)

+++ core/embed/io/display/ltdc_dsi/display_driver.c
@@ -160,12 +160,10 @@ static bool display_dsi_init(display_driver_t *drv) {
   drv->DSIVidCfg.PacketSize = HACT;
   drv->DSIVidCfg.NumberOfChunks = 0;  // No chunks in burst mode
   drv->DSIVidCfg.NullPacketSize = 0;  // No null packet in burst mode
-  drv->DSIVidCfg.HorizontalSyncActive =
-      (uint32_t)(HSYNC * LANE_BYTE_2_PIXEL_CLK_RATIO);
-  drv->DSIVidCfg.HorizontalBackPorch =
-      (uint32_t)(HBP * LANE_BYTE_2_PIXEL_CLK_RATIO);
+  drv->DSIVidCfg.HorizontalSyncActive = HSYNC * LANE_BYTE_2_PIXEL_CLK_RATIO;
+  drv->DSIVidCfg.HorizontalBackPorch = HBP * LANE_BYTE_2_PIXEL_CLK_RATIO;
   drv->DSIVidCfg.HorizontalLine =
-      (uint32_t)((HSYNC + HBP + HACT + HFP) * LANE_BYTE_2_PIXEL_CLK_RATIO);
+      (HSYNC + HBP + HACT + HFP) * LANE_BYTE_2_PIXEL_CLK_RATIO;
   drv->DSIVidCfg.VerticalSyncActive = VSYNC;
   drv->DSIVidCfg.VerticalBackPorch = VBP;
   drv->DSIVidCfg.VerticalFrontPorch = VFP;
